Refactor `diff` into `patches/` folder
