+++ lib/minitest/difftastic.rb
@@ -1,8 +1,15 @@
 # frozen_string_literal: true
 
-require_relative "difftastic/version"
+require "difftastic"
 
-module Minitest
-  module Difftastic
-  end
+require "minitest/difftastic/version"
+
+module Minitest::Difftastic
+  DEFAULT_DIFFER = ::Difftastic::Differ.new(
+    color: :always,
+    tab_width: 2,
+    syntax_highlight: :off,
+    left_label: "Expected",
+    right_label: "Actual"
+  )
 end
