+++ lib/minitest/difftastic_plugin.rb
@@ -1,25 +1,5 @@
 # frozen_string_literal: true
 
-require "difftastic"
+require "minitest/difftastic"
 
-# TODO: use refinements
-module Minitest
-  module Assertions
-    DIFFER = ::Difftastic::Differ.new(
-      color: :always,
-      tab_width: 2,
-      syntax_highlight: :off,
-      left_label: "Expected",
-      right_label: "Actual"
-    )
-
-    alias diff_original diff
-
-    def diff(exp, act)
-      DIFFER.diff_objects(exp, act)
-    rescue StandardError => e
-      puts "Minitest::Difftastic error: #{e.inspect} (#{e.backtrace[0]})"
-      diff_original(exp, act)
-    end
-  end
-end
+require "minitest/difftastic/patches/diff"
